+++ src/main/java/com/accenture/granja/repository/CompraRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.accenture.granja.model.Compra;
 
-public interface CompraRepository extends JpaRepository<Compra, Long>{
+public interface CompraRepository extends JpaRepository<Compra, Integer>{
 
 }
